code rearrange
