+++ main.py
@@ -8,7 +8,6 @@ from src.ground import *
 from src.cloud import *
 from src.scoreboard import *
 from src.item import *
-from src.life_item import *
 from src.heart import *
 from db_interface import InterfDB
 
@@ -98,13 +97,13 @@ def gameplay():
     pteras = pygame.sprite.Group()
     clouds = pygame.sprite.Group()
     last_obstacle = pygame.sprite.Group()
-    items = pygame.sprite.Group()
+    shield_items = pygame.sprite.Group()
     life_items = pygame.sprite.Group()
 
     Cactus.containers = cacti
     Ptera.containers = pteras
     Cloud.containers = clouds
-    Item.containers = items
+    ShieldItem.containers = shield_items
     LifeItem.containers = life_items
 
     retbutton_image, retbutton_rect = load_image('replay_button.png', 35, 31, -1)
@@ -211,20 +210,20 @@ def gameplay():
                             playerDino.collision_immune = False
 
                 if not playerDino.isSuper:
-                    for i in items:
-                        i.movement[0] = -1 * gamespeed
-                        if pygame.sprite.collide_mask(playerDino, i):
+                    for s in shield_items:
+                        s.movement[0] = -1 * gamespeed
+                        if pygame.sprite.collide_mask(playerDino, s):
                             playerDino.collision_immune = True
                             playerDino.isSuper = True
-                            i.kill()
+                            s.kill()
                             item_time = pygame.time.get_ticks()
                 else:
-                    for i in items:
-                        i.movement[0] = -1 * gamespeed
-                        if pygame.sprite.collide_mask(playerDino, i):
+                    for s in shield_items:
+                        s.movement[0] = -1 * gamespeed
+                        if pygame.sprite.collide_mask(playerDino, s):
                             playerDino.collision_immune = True
                             playerDino.isSuper = True
-                            i.kill()
+                            s.kill()
                             item_time = pygame.time.get_ticks()
 
                     if pygame.time.get_ticks() - item_time > 2000:
@@ -256,11 +255,11 @@ def gameplay():
                 if len(clouds) < 5 and random.randrange(0, 300) == 10:
                     Cloud(width, random.randrange(height / 5, height / 2))
 
-                if len(items) == 0 and random.randrange(0, 200) == 10 and counter > 300:
+                if len(shield_items) == 0 and random.randrange(0, 200) == 10 and counter > 300:
                     for l in last_obstacle:
                         if l.rect.right < width * 0.8:
                             last_obstacle.empty()
-                            last_obstacle.add(Item(gamespeed, 46, 40))
+                            last_obstacle.add(ShieldItem(gamespeed, 46, 40))
 
                 if len(life_items) == 0 and random.randrange(0, 300) == 10 and counter > 400:
                     for l in last_obstacle:
@@ -272,7 +271,7 @@ def gameplay():
                 cacti.update()
                 pteras.update()
                 clouds.update()
-                items.update()
+                shield_items.update()
                 life_items.update()
                 new_ground.update()
                 scb.update(playerDino.score)
@@ -290,7 +289,7 @@ def gameplay():
                         screen.blit(HI_image, HI_rect)
                     cacti.draw(screen)
                     pteras.draw(screen)
-                    items.draw(screen)
+                    shield_items.draw(screen)
                     life_items.draw(screen)
                     playerDino.draw()
                     resized_screen.blit(
